Invoke podman/docker run afresh in a container restart instead of "container stop/start" (#2)
The current way the container service works is by stopping and starting the container to utilize its previous state. Use fresh `podman/docker run` invocation in container restart instead of `container start` since latter fails in a variety of cases where there have been even the slightest of changes in the system environment:

- use `podman/docker run` to start the ybox container for all cases in the service file
  * updated the service template for this with `CONTAINER_RUN_ARGS` environment variable in the env file that has all the arguments required for the podman/docker run execution
  * skip some env arguments to the run execution like `REQUIRED_PKGS` which are for installing packages only in the initial create and can be quite large, so also reduces the size of `CONTAINER_RUN_ARGS` substantially
  * `ybox-control stop` now also has a `--rm` argument to also remove the container which is now used by the service stop
- use host mounted directories for /usr,/var,... for the non-shared root case too
  * a similar sequence as used for `shared_root` of running a temporary "copy container" is now used for the non-shared root case too; the destination directory is deliberately fixed and not configurable to simplify things a bit
  * renamed `shared_root_dirs` property in distro.ini to `mount_root_dirs` to reflect the fact that it is not specific to the `shared_root` case
  * delete the non-shared root directory on `ybox-destroy` using `podman unshare`; also added removing of the `configs` and `ybox-scripts` directories and optionally also the home and entire container directory with the `-D`/`--delete-files` option to `ybox-destroy`

Few other changes:
- fixed locale-gen for Arch/CachyOS where locale-gen exit value may still be 0 in case a successful one was generated (e.g. en_US may have succeeded by en_IN may have failed), so always reinstall glibc to get all the locales
- removed some non-required packages (like `bat`) in distro.ini of distros

+++ src/ybox/conf/resources/entrypoint-root.sh
@@ -9,11 +9,13 @@ SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 source "$SCRIPT_DIR/entrypoint-common.sh"
 
 export HOME=/root
-echo_color "$fg_cyan" "Copying prime-run, run-in-dir and run-user-bash-cmd" >> $status_file
-cp -af "$SCRIPT_DIR/prime-run" /usr/local/bin/prime-run
-cp -af "$SCRIPT_DIR/run-in-dir" /usr/local/bin/run-in-dir
-cp -af "$SCRIPT_DIR/run-user-bash-cmd" /usr/local/bin/run-user-bash-cmd
-chmod 0755 /usr/local/bin/prime-run /usr/local/bin/run-in-dir /usr/local/bin/run-user-bash-cmd
+if [ ! -e "$SCRIPT_DIR/ybox-init.done" ]; then
+  echo_color "$fg_cyan" "Copying prime-run, run-in-dir and run-user-bash-cmd" >> $status_file
+  cp -af "$SCRIPT_DIR/prime-run" /usr/local/bin/prime-run
+  cp -af "$SCRIPT_DIR/run-in-dir" /usr/local/bin/run-in-dir
+  cp -af "$SCRIPT_DIR/run-user-bash-cmd" /usr/local/bin/run-user-bash-cmd
+  chmod 0755 /usr/local/bin/prime-run /usr/local/bin/run-in-dir /usr/local/bin/run-user-bash-cmd
+fi
 
 # invoke the NVIDIA setup script if present
 if [ -r "$SCRIPT_DIR/nvidia-setup.sh" ]; then
